Make all the tests pass.
