+++ paxy/compiler/assembler.py
@@ -18,7 +18,10 @@ from paxy.compiler.ir import (
     COND_JUMP_OPS,
     UNCOND_JUMP_FIXED,
 )
-from paxy.compiler.opcoerce import normalize_push_null_for_calls_312_seq
+from paxy.compiler.twelve import (
+    normalize_push_null_for_calls_312_seq,
+    try_func_to_code_with_endfor_fix,
+)
 
 
 # What the resolver returns (only real bytecode items)
@@ -244,12 +247,14 @@ class Assembler:
 
         if sys.version_info >= (3, 13):
             maker = Instr("MAKE_FUNCTION", lineno=func.lineno)
+            func_code = bc_func.to_code()
         else:
             maker = Instr("MAKE_FUNCTION", 0, lineno=func.lineno)
+            func_code = try_func_to_code_with_endfor_fix(bc_func)
 
         # 7) Emit loader sequence
         return [
-            Instr("LOAD_CONST", bc_func.to_code(), lineno=func.lineno),
+            Instr("LOAD_CONST", func_code, lineno=func.lineno),
             maker,
             Instr("STORE_NAME", func.name, lineno=func.lineno),
         ]
@@ -499,8 +504,8 @@ class Assembler:
         # 4) Loop end + cleanup (tests want POP_TOP present)
         out.append(l_end)
         out.append(Instr("END_FOR", lineno=it.lineno))
-        if sys.version_info >= (3, 13):
-            out.append(Instr("POP_TOP", lineno=it.lineno))
+        # if sys.version_info >= (3, 13):
+        out.append(Instr("POP_TOP", lineno=it.lineno))
 
         return out
